backend/tests/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
